Look at 'named as' field in reference for source name
The official site lists members by a fuller version of their name than
is commonly used, so look for a "named as" field on the reference to
compare with instead.

(This will fail if there is more than one such "named as", e.g. across
separate references for the same statement, but we can fix that kui it
happens)